Alsia/WMDP-163 Add CD to the screener (#67)
Ticket
https://wicmtdp.atlassian.net/browse/WMDP-164
https://wicmtdp.atlassian.net/browse/WMDP-163

Changes
What was added, updated, or removed in this PR.

Context for reviewers
CD should:
Build the image
Store the image in ECR
Run aws ecs update-service --cluster --service --force-new-deployment after we confirm the image is in ECR in order to force the update

There is a commented-out iam.tf which shows the roles and permissions created as a part of getting this action running. These will live in the wic-mt-demo repo.