update service name in workflow (#96)
Ticket
https://wicmtdp.atlassian.net/browse/WMDP-40

Changes
Update the service name in the deployment workflow
Context for reviewers
As a part of the changes made here the ecs service name was updated for clarity. This change resulted in breaking the deployment job because it is now referencing a service that is no longer ACTIVE and in use.