[Post-Demo] Migrate Terraform code from project repo (#186)
Ticket
https://wicmtdp.atlassian.net/browse/PRP-116

Changes
infra/eligibility_screener path added with the infrastructure for the eligibility screener repo
Context for reviewers
This ticket is for a migration of the terraform out of the project repo and into the Eligibility Screener repo. The end result should be that we should be able to manage AWS via terraform for the Eligibility Screener.

This ticket should not include any major refactoring of the terraform structure.